Rename the con file to config
