+++ app/db/courses.py
@@ -1,4 +1,4 @@
-from .con import get_db_connection
+from .connector import get_db_connection
 from mysql.connector import Error
 
 class Course:
